게시글 권한추가하기

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -8,6 +8,8 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding.blog._core.errors.exception.Exception403;
+import shop.mtcoding.blog._core.errors.exception.Exception404;
 import shop.mtcoding.blog.user.User;
 
 import java.util.List;
@@ -21,6 +23,13 @@ public class BoardController {
 
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable Integer id, BoardRequest.UpdateDTO reqDTO){
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글 수정 권한 없음");
+        }
+
         boardRepository.updateById(id, reqDTO.getTitle(), reqDTO.getContent());
         return "redirect:/board/" + id;
     }
@@ -28,12 +37,24 @@ public class BoardController {
     @GetMapping("/board/{id}/update-form")
     public String updateForm(@PathVariable Integer id, HttpServletRequest request){
         Board board = boardRepository.findById(id);
+
+        if (board == null){
+            throw new Exception404("해당 게시글 찾을 수 없음");
+        }
+
         request.setAttribute("board", board);
         return "board/update-form";
     }
 
     @PostMapping("/board/{id}/delete")
     public String delete(@PathVariable Integer id){
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글 삭제 권한 없음");
+        }
+
         boardRepository.deleteById(id);
         return "redirect:/";
     }
@@ -61,7 +82,20 @@ public class BoardController {
 
     @GetMapping("/board/{id}")
     public String detail(@PathVariable Integer id, HttpServletRequest request) {
+
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findByIdJoinUser(id);
+
+        // 게시글의 주인이 아니라면 -> 로그인을 했으면 ->
+        // 로그인 한 유저가 = 게시글 쓴 사람과 일치하면 ->
+        // 주인이 맞다 -> else면 아니다
+        boolean isOwner = false;
+        if (sessionUser != null){
+            if (sessionUser.getId() == board.getUser().getId()){
+                isOwner = true;
+            }
+        }
+        request.setAttribute("isOwner", isOwner);
         request.setAttribute("board", board);
         return "board/detail";
     }
